Moved the "players+" command to the DB module + improvmentss
